Split parser atom forms
